requirements.txt
python-dotenv
fastapi[standard]
streamlit
litellm
